refactor: rename legal component files to follow PascalCase convention
- zh-cn-privacy.mdx → ZhCnPrivacy.mdx
- en-privacy.mdx → EnPrivacy.mdx
- zh-cn-agreement.mdx → ZhCnAgreement.mdx
- en-agreement.mdx → EnAgreement.mdx
- Update import statements in privacy/page.tsx and agreement/page.tsx

As these are now components in the shared components directory,
they should follow PascalCase naming convention.
