+++ src/cook-web/src/components/legals/EnAgreement.mdx
