+++ src/cook-web/src/components/legals/EnPrivacy.mdx
