+++ src/cook-web/src/components/legals/ZhCnAgreement.mdx
